refactor(refactor): rename www to http namespace
