+++ tests/http/ServerSocketTest.cpp
@@ -1,9 +1,9 @@
 
 #include <gtest/gtest.h>
 
-#include "www/ServerSocket.hpp"
+#include "http/ServerSocket.hpp"
 
-namespace www {
+namespace http {
 
 class ServerSocketTest : public ::testing::Test {
 	protected:
@@ -14,4 +14,4 @@ TEST_F(ServerSocketTest, DefaultConstructor) {
 	ASSERT_NO_THROW(ServerSocket new_instance;);
 }
 
-} /* namespace www */
+}  // namespace http
